+++ auth_test.go
@@ -60,7 +60,8 @@ func TestAuth_GrantIsIdempotent(t *testing.T) {
 	// Lets assume day already has permissions ot open the pod bay doors we want
 	// to make sure we will not append the same permissions multiple times.
 	mem.On("Get", "joe.permissions.dave").Return(`["open_pod_bay_doors","foo.bar"]`, true, nil)
-	mem.On("Set", "joe.permissions.dave", `["open_pod_bay_doors","foo.bar"]`).Return(nil)
+
+	// We should not see a call to mem.Set(…) since no permissions need to be updated
 
 	ok, err := auth.Grant("open_pod_bay_doors", "dave")
 	require.NoError(t, err)
@@ -84,6 +85,70 @@ func TestAuth_GrantWiderScope(t *testing.T) {
 	assert.True(t, ok)
 
 	mem.AssertExpectations(t)
+}
+
+func TestAuth_GrantSmallerScope(t *testing.T) {
+	logger := zaptest.NewLogger(t)
+	mem := new(memoryMock)
+	auth := NewAuth(logger, mem)
+
+	mem.On("Get", "joe.permissions.fgrosse").Return(`["foo", "test"]`, true, nil)
+
+	// We should not see a call to mem.Set(…) since no permissions were actually added
+
+	ok, err := auth.Grant("foo.bar.baz", "fgrosse")
+	require.NoError(t, err)
+	assert.False(t, ok)
+
+	mem.AssertExpectations(t)
+}
+
+func TestAuth_Revoke(t *testing.T) {
+	logger := zaptest.NewLogger(t)
+	mem := new(memoryMock)
+	auth := NewAuth(logger, mem)
+
+	mem.On("Get", "joe.permissions.fgrosse").Return(`["test"]`, true, nil)
+	ok, err := auth.Revoke("foo.bar", "fgrosse")
+	assert.NoError(t, err)
+	assert.False(t, ok)
+	mem.AssertExpectations(t)
+
+	mem = new(memoryMock)
+	auth = NewAuth(logger, mem)
+
+	mem.On("Get", "joe.permissions.fgrosse").Return(`["foo.bar", "test"]`, true, nil)
+	mem.On("Set", "joe.permissions.fgrosse", `["test"]`).Return(nil)
+
+	ok, err = auth.Revoke("foo.bar", "fgrosse")
+	assert.NoError(t, err)
+	assert.True(t, ok)
+
+	// TODO: what if we revoke "foo.bar" and the user has the wider "foo" scope?
+
+	mem.AssertExpectations(t)
+}
+
+func TestAuth_RevokeWiderScope(t *testing.T) {
+	logger := zaptest.NewLogger(t)
+	mem := new(memoryMock)
+	auth := NewAuth(logger, mem)
+
+	mem.On("Get", "joe.permissions.fgrosse").Return(`["foo"]`, true, nil)
+
+	ok, err := auth.Revoke("foo.bar", "fgrosse")
+	assert.EqualError(t, err, `cannot revoke scope "foo.bar" because the user still has the more general scope "foo"`)
+	assert.False(t, ok)
+}
+
+func TestAuth_RevokeEmptyScope(t *testing.T) {
+	logger := zaptest.NewLogger(t)
+	mem := new(memoryMock)
+	auth := NewAuth(logger, mem)
+
+	ok, err := auth.Revoke("", "fgrosse")
+	assert.EqualError(t, err, "scope cannot be empty")
+	assert.False(t, ok)
 }
 
 func TestAuth_CheckPermission_Errors(t *testing.T) {