+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **THIS SOFTWARE IS STILL IN ALPHA AND THERE ARE NO GUARANTEES REGARDING API STABILITY YET.**
 
 ## [Unreleased]
-- Make Auth.Grant(…) idempotent
+- Make Auth.Grant(…) idempotent and do not unnecessarily add smaller scopes
 - Support extending permissions via Auth.Grant(…)
 - Add boolean return value to Auth.Grant(…) to indicate if a new permission was granted
+- Add Auth.Revoke(…) to remove permissions
 
 ## [v0.7.0] - 2019-04-18
 - Add ReceiveMessageEvent.Data field to allow using the underlying message type of the adapters